requirements.txt
streamlit>=1.32
fastapi>=0.110
uvicorn[standard]>=0.29
sqlalchemy>=2.0
psycopg2-binary>=2.9
geopandas>=0.14
pandas>=2.2
numpy>=1.26
plotly>=5.19
folium>=0.16
streamlit-folium>=0.18
requests>=2.31
pyyaml>=6.0
scipy>=1.12
scikit-learn>=1.4
apscheduler>=3.10
pytest>=8.0
httpx>=0.27
